example/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
